+++ tectonic/xetex-XeTeXLayoutInterface.cpp
@@ -43,7 +43,7 @@ authorization from the copyright holders.
 #include <hb-icu.h>
 #include <hb-ot.h>
 
-#include "XeTeX_web.h"
+#include "xetex-web.h"
 
 #include "xetex-XeTeXLayoutInterface.h"
 #include "xetex-XeTeXFontInst.h"
